Syntax fixing
If the syntax is correct now the next step would be to update algo.rb to
make full use of the nodes.
We might want to rename the node.rb file too. At it's core it has a
MinMax logic.
Maybe we can create an abstract node class from wich the specific nodes
inherit.